add ReadOnlyEventStore interface
resolves: https://github.com/prooph/event-store/issues/250
